// aynakart/src/app/product.interface.ts
// Here ProductDetails refer to Product Entity
export interface ProductDetails {
    productId: number;
    sellerName: string;
    productName: string;
    brandName: string;
    categoryId: number;
    stockUnits: number;
    unitPrice: number;
    listedDate: string;
    imageUrl: string;
  }
  